+++ openrouteservice-api-tests/src/test/java/heigit/ors/v2/services/serviceSettings/MatrixServiceSettings.java
@@ -16,8 +16,8 @@ package heigit.ors.v2.services.serviceSettings;
 import heigit.ors.v2.services.config.AppConfig;
 
 public class MatrixServiceSettings {
-    private static int maximumLocations = 100;
-    private static int maximumLocationsFlexible = 25;
+    private static int maximumRoutes = 100;
+    private static int maximumRoutesFlexible = 25;
     private static int maximumVisitedNodes = 100000;
     private static double maximumSearchRadius = 2000;
     private static boolean allowResolveLocations = true;
@@ -30,10 +30,10 @@ public class MatrixServiceSettings {
             enabled = Boolean.parseBoolean(value);
         value = AppConfig.Global().getServiceParameter("matrix", "maximum_locations");
         if (value != null)
-            maximumLocations = Math.max(1, Integer.parseInt(value));
+            maximumRoutes = Math.max(1, Integer.parseInt(value));
         value = AppConfig.Global().getServiceParameter("matrix", "maximum_locations_flexible");
         if (value != null)
-            maximumLocationsFlexible = Math.max(1, Integer.parseInt(value));
+            maximumRoutesFlexible = Math.max(1, Integer.parseInt(value));
         value = AppConfig.Global().getServiceParameter("matrix", "maximum_search_radius");
         if (value != null)
             maximumSearchRadius = Math.max(1, Double.parseDouble(value));
@@ -60,8 +60,8 @@ public class MatrixServiceSettings {
         return maximumVisitedNodes;
     }
 
-    public static int getMaximumLocations(boolean flexible) {
-        return (flexible ? maximumLocationsFlexible : maximumLocations);
+    public static int getMaximumRoutes(boolean flexible) {
+        return (flexible ? maximumRoutesFlexible : maximumRoutes);
     }
 
     public static double getMaximumSearchRadius() {